add api boxType sellPrice pullPrice

+++ domain/src/main/java/com/paya/domain/models/repo/CashManagerRepModel.kt
@@ -3,4 +3,4 @@ package com.paya.domain.models.repo
 data class CashManagerRepModel(var pricingCash: List<PricingCash>)
 
 
-data class PricingCash(val min:Int,val max:Int)
+data class PricingCash(val min:Long,val max:Long)
